Tidy up test code
tintoy/msbuild-project-tools-server#100
